dev-requirements.txt
-r requirements.txt

black==20.8b1
pytest-mock
